fix: 로직분리
