+++ src/src/components/my-fund/approve-withdrawal/WithdrawalApproval.jsx
@@ -1,15 +1,14 @@
 import { Suspense } from "react";
-import WithdrawalFundInfo from "@/components/my-fund/approve-withdrawal/WithdrawalFundInfo.jsx";
-import InfiniteWithdrawInfoLoader from "@/components/my-fund/approve-withdrawal/InfiniteWithdrawInfo.loader.jsx";
-
+import InfiniteWithdrawalFundInfo from "@/components/my-fund/approve-withdrawal/InfiniteWithdrawalFundInfo.jsx";
+import InfiniteWithdrawalFundInfoLoader from "./InfiniteWithdrawaIFundInfo.loader.jsx";
 /**
  * My펀딩 출금승인 tab 해당내용
  */
 
 function WithdrawalApproval() {
   return (
-    <Suspense fallback={<InfiniteWithdrawInfoLoader />}>
-      <WithdrawalFundInfo />
+    <Suspense fallback={<InfiniteWithdrawalFundInfoLoader />}>
+      <InfiniteWithdrawalFundInfo />
     </Suspense>
   );
 }
